move graph to appropriate package (#174)
This PR moves `graph` package under `metadata` to the appropriate package in `evaluators`.
