+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/evaluators/AbstractEvaluator.java
@@ -27,9 +27,9 @@ package edu.ucr.cs.riple.core.evaluators;
 import com.google.common.collect.ImmutableSet;
 import edu.ucr.cs.riple.core.Config;
 import edu.ucr.cs.riple.core.Report;
-import edu.ucr.cs.riple.core.evaluators.graphprocessor.ConflictGraphProcessor;
+import edu.ucr.cs.riple.core.evaluators.graph.ConflictGraph;
+import edu.ucr.cs.riple.core.evaluators.graph.processors.ConflictGraphProcessor;
 import edu.ucr.cs.riple.core.evaluators.suppliers.Supplier;
-import edu.ucr.cs.riple.core.metadata.graph.ConflictGraph;
 import edu.ucr.cs.riple.core.metadata.index.Fix;
 
 /**
