+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/evaluators/BasicEvaluator.java
@@ -27,8 +27,8 @@ package edu.ucr.cs.riple.core.evaluators;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Sets;
 import edu.ucr.cs.riple.core.Report;
+import edu.ucr.cs.riple.core.evaluators.graph.Node;
 import edu.ucr.cs.riple.core.evaluators.suppliers.Supplier;
-import edu.ucr.cs.riple.core.metadata.graph.Node;
 import edu.ucr.cs.riple.core.metadata.index.Fix;
 
 /**
