+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/evaluators/graph/Node.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package edu.ucr.cs.riple.core.metadata.graph;
+package edu.ucr.cs.riple.core.evaluators.graph;
 
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Sets;
