+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/evaluators/suppliers/TargetModuleSupplier.java
@@ -28,10 +28,10 @@ import com.google.common.collect.ImmutableSet;
 import edu.ucr.cs.riple.core.Config;
 import edu.ucr.cs.riple.core.cache.TargetModuleCache;
 import edu.ucr.cs.riple.core.cache.downstream.DownstreamImpactCache;
-import edu.ucr.cs.riple.core.evaluators.graphprocessor.CompilerRunner;
-import edu.ucr.cs.riple.core.evaluators.graphprocessor.ConflictGraphProcessor;
-import edu.ucr.cs.riple.core.evaluators.graphprocessor.ParallelConflictGraphProcessor;
-import edu.ucr.cs.riple.core.evaluators.graphprocessor.SequentialConflictGraphProcessor;
+import edu.ucr.cs.riple.core.evaluators.graph.processors.CompilerRunner;
+import edu.ucr.cs.riple.core.evaluators.graph.processors.ConflictGraphProcessor;
+import edu.ucr.cs.riple.core.evaluators.graph.processors.ParallelConflictGraphProcessor;
+import edu.ucr.cs.riple.core.evaluators.graph.processors.SequentialConflictGraphProcessor;
 import edu.ucr.cs.riple.core.injectors.AnnotationInjector;
 import edu.ucr.cs.riple.core.injectors.PhysicalInjector;
 import edu.ucr.cs.riple.core.metadata.method.MethodRegistry;
