+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/metadata/trackers/generatedcode/LombokTracker.java
@@ -24,6 +24,7 @@
 
 package edu.ucr.cs.riple.core.metadata.trackers.generatedcode;
 
+import edu.ucr.cs.riple.core.evaluators.graph.processors.ParallelConflictGraphProcessor;
 import edu.ucr.cs.riple.core.metadata.method.MethodRegistry;
 import edu.ucr.cs.riple.core.metadata.trackers.MethodRegionTracker;
 import edu.ucr.cs.riple.core.metadata.trackers.Region;
@@ -40,8 +41,8 @@ import java.util.stream.Stream;
  * {@code @Nullable} annotation on fields to getter methods, therefore, extends the set of
  * potentially impacted regions to all callers of that method as well. This tracker, will include
  * all callers of any method region in lombok generated code. This will guarantee that {@link
- * edu.ucr.cs.riple.core.evaluators.graphprocessor.ParallelConflictGraphProcessor} will catch any
- * triggered errors by an annotation including all copied annotations by lombok as well.
+ * ParallelConflictGraphProcessor} will catch any triggered errors by an annotation including all
+ * copied annotations by lombok as well.
  */
 public class LombokTracker implements GeneratedRegionTracker {
